Add full yarn support
- It closes issue #4.

With yarn there are several problems. A package can not be installed if
it's not referenced in the package.json file. It makes no sense here
because we use submodules and symlinks in node_modules. It's necessary
to have a backup of the package.json file in order to restore the
content after all yarn add calls. We cannot use 'install', we can use
only 'add' and it's a contraint of yarn.

A second problem is these calls because it runs the main postinstall
script again. Then an environment variable is used in order to skip
startcraft when it's called fromn startcraft.