Generate or use a full package-lock for consistency in installations
Note that it's faster by this way. The consistency is preserved when the
installation is done by the postinstall hook.

The `full-package-lock.json` file must be commited in the repository. If
you want to update the versions of all dependencies, then remove this
file, a new one will be created with all the updates.